# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Settlement API
  description: |-
    日本企業の経費申請・承認フローを想定した REST API。
    `docs/openapi.yaml` を正式契約とし、Swagger UI と Prism Mock Server が同じ定義を利用する。
  version: 0.6.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:4010
    description: Prism Mock Server
...
paths:
...
  /api/expense-applications/{applicationId}/items/{itemId}/receipt:
    get:
      tags:
        - Receipts
      summary: 領収書 metadata 取得
      description: 認可済み user に active receipt の公開可能 metadata を返す。Storage key/path は返さない。
      operationId: getReceiptMetadata
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReceiptApplicationId'
        - $ref: '#/components/parameters/ExpenseItemId'
      responses:
        '200':
          description: Metadata 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptFileApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Receipts
      summary: 領収書 upload / replace
      description: 申請者本人が DRAFT / RETURNED の明細へ JPEG/PNG/PDF を upload する。既存 active receipt がある場合は検査完了後に replace する。
      operationId: putReceipt
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReceiptApplicationId'
        - $ref: '#/components/parameters/ExpenseItemId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG、PNG、PDF。最大 10 MiB。
      responses:
        '200':
          description: Upload / replace 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptFileApiResponse'
        '400':
          $ref: '#/components/responses/InvalidFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenOrCsrfInvalid'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/FileTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/MalwareDetected'
        '503':
          $ref: '#/components/responses/FileServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Receipts
      summary: 領収書 delete
      description: 申請者本人が DRAFT / RETURNED の active receipt を利用者から参照不可にし、object cleanup を登録する。
      operationId: deleteReceipt
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReceiptApplicationId'
        - $ref: '#/components/parameters/ExpenseItemId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Delete 受付成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenOrCsrfInvalid'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/FileServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/expense-applications/{applicationId}/items/{itemId}/receipt/content:
    get:
      tags:
        - Receipts
      summary: 領収書 preview / download
      description: 認可後、active receipt を private storage から streaming する。Binary response は ApiResponse wrapper を使用しない。
      operationId: getReceiptContent
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReceiptApplicationId'
        - $ref: '#/components/parameters/ExpenseItemId'
        - name: disposition
          in: query
          required: false
          schema:
            type: string
            enum:
              - inline
              - attachment
            default: inline
      responses:
        '200':
          description: File content。Content-Disposition、X-Content-Type-Options、Cache-Control を設定する。
          headers:
            Content-Disposition:
              description: Sanitized filename を持つ inline または attachment。
              schema:
                type: string
            X-Content-Type-Options:
              description: 常に nosniff。
              schema:
                type: string
                enum: [nosniff]
            Cache-Control:
              description: 常に private, no-store。
              schema:
                type: string
                example: private, no-store
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/FileServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Session JDBC が管理する opaque server-side session ID。Browser JavaScript から読み取らない。
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
      description: "/api/auth/csrf が返す token。POST / PUT / PATCH / DELETE で必須。"
...
    ReceiptApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 領収書が属する経費申請 ID
    ExpenseItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 経費明細 ID。applicationId との親子関係を検証する。
  responses:
    BadRequest:
      description: リクエスト不正または業務ルール違反
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              summary: Validation error
              value:
                success: false
                code: VALIDATION_ERROR
                message: 入力内容に誤りがあります。
                details:
                  - field: title
                    message: 空白は許可されていません
                path: /api/expense-applications
                timestamp: '2026-07-12T10:00:00'
            businessRuleError:
              summary: 業務ルール違反
              value:
                success: false
                code: BAD_REQUEST
                message: 申請中の経費申請のみ承認・差戻しできます。
                path: /api/expense-applications/10/approve
                timestamp: '2026-07-12T10:00:00'
    Unauthorized:
      description: 未認証または認証失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: UNAUTHORIZED
            message: 認証が必要です。
            path: /api/expense-applications
            timestamp: '2026-07-12T10:00:00'
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FORBIDDEN
            message: 他のユーザーの経費申請は操作できません。
            path: /api/expense-applications/10
            timestamp: '2026-07-12T10:00:00'
...
    ForbiddenOrCsrfInvalid:
      description: Role/ownership/status の権限不足、または CSRF token 不足・不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                success: false
                code: FORBIDDEN
                message: この領収書は操作できません。
                path: /api/expense-applications/10/items/100/receipt
                timestamp: '2026-07-23T10:00:00'
            csrfInvalid:
              value:
                success: false
                code: CSRF_INVALID
                message: セキュリティトークンが無効です。再読み込みして再試行してください。
                path: /api/expense-applications/10/items/100/receipt
                timestamp: '2026-07-23T10:00:00'
    NotFound:
      description: 対象データなし
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: NOT_FOUND
            message: 経費申請が見つかりません。
            path: /api/expense-applications/999
            timestamp: '2026-07-12T10:00:00'
    InvalidFile:
      description: Empty file、file name、extension/Content-Type/signature の不整合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INVALID_FILE
            message: ファイルの内容を確認してください。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    Conflict:
      description: Concurrent replace/delete または receipt state 競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: CONFLICT
            message: 領収書が更新されています。最新の状態を確認してください。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    FileTooLarge:
      description: Upload file が 10 MiB を超過
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FILE_TOO_LARGE
            message: ファイルは10 MiB以下にしてください。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    UnsupportedMediaType:
      description: JPEG、PNG、PDF 以外
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: UNSUPPORTED_MEDIA_TYPE
            message: JPEG、PNG、PDF形式のファイルを選択してください。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    MalwareDetected:
      description: Malware scanner が危険な file と判定
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: MALWARE_DETECTED
            message: 安全性を確認できないため、このファイルは登録できません。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    FileServiceUnavailable:
      description: Private storage または malware scanner の一時障害
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FILE_SERVICE_UNAVAILABLE
            message: ファイルサービスを一時的に利用できません。時間をおいて再試行してください。
            path: /api/expense-applications/10/items/100/receipt
            timestamp: '2026-07-23T10:00:00'
    InternalServerError:
      description: 予期しないシステムエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INTERNAL_SERVER_ERROR
            message: システムエラーが発生しました。
            path: /api/expense-applications
            timestamp: '2026-07-12T10:00:00'
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - code
        - message
        - path
        - timestamp
      properties:
        success:
          type: boolean
          enum: [false]
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: 入力内容に誤りがあります。
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
        path:
          type: string
          example: /api/expense-applications
        timestamp:
          type: string
          description: ISO-8601 local date-time。アプリケーションの timezone は Asia/Tokyo。
          example: '2026-07-12T10:00:00'
    ValidationErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: title
        message:
          type: string
          example: 空白は許可されていません
    ApiResponseBase:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          description: ISO-8601 local date-time。アプリケーションの timezone は Asia/Tokyo。
          example: '2026-07-12T10:00:00'
...
    ReceiptFileApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ReceiptFileResponse'
      example:
        success: true
        data:
          id: 500
          originalFileName: tokyo-station-receipt.pdf
          contentType: application/pdf
          sizeBytes: 245760
          sha256Checksum: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
          uploadedAt: '2026-07-23T10:00:00'
          previewAvailable: true
        timestamp: '2026-07-23T10:00:00'
...
    VoidApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: object
              nullable: true
      example:
        success: true
        message: 経費申請を削除しました。
        timestamp: '2026-07-12T10:00:00'
...
    ReceiptFileResponse:
      type: object
      required:
        - id
        - originalFileName
        - contentType
        - sizeBytes
        - sha256Checksum
        - uploadedAt
        - previewAvailable
      properties:
        id:
          type: integer
          format: int64
        originalFileName:
          type: string
          maxLength: 255
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - application/pdf
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          maximum: 10485760
        sha256Checksum:
          type: string
          pattern: '^[0-9a-f]{64}$'
        uploadedAt:
          type: string
          format: date-time
          description: ISO-8601 local date-time
        previewAvailable:
          type: boolean
...